# openapi-spec/build/openapi.yaml
openapi: 3.0.2
info:
  title: Fullstack Portfolio API
  version: 1.0.0
  license:
    name: private
  description: FastAPI of Fullstack Portfolio
servers:
  - description: local
    url: http://localhost:8000/api/v1
  - description: local-fastapi
    url: http://localhost
paths:
...
  /machines/:
    get:
      summary: fetches machines
      operationId: fetchMachines
      responses:
        '201':
          description: A list of machines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/machines'
        '404':
          description: Not found
        default:
          $ref: '#/components/responses/DefaultError'
      tags:
        - Machines
...
components:
  securitySchemes:
    userAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    contactAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenericError:
      type: object
      properties:
        details:
          type: string
...
    machine:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
          example: 2021-01-30T08:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2021-01-30T08:30:00Z
        unit_installed_at:
          type: string
          format: date-time
          example: 2021-01-30T08:30:00Z
        oem_name:
          type: string
        model:
          type: string
        make:
          type: string
        equipment_id:
          type: string
        serial_number:
          type: string
        pin:
          type: string
    machines:
      type: array
      items:
        $ref: '#/components/schemas/machine'
...
  responses:
    DefaultError:
      description: Default Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
...
security:
  - userAccessToken: []